Adds debugging GL checks and limits VBOs to shorts.
WebGL / GLES2.0 can only support up to shorts for VBO element buffers
unless there is an extension. However, WebGL in particular is rather
strict. We should query for this case and use a workaround when
necessary.

+++ include/apsis/model/mesh.h
@@ -21,7 +21,7 @@ namespace Apsis {
       Mesh(std::vector<glm::vec3>& vertices,
            std::vector<glm::vec3>& normals,
            std::vector<glm::vec2>& textureCoords,
-           std::vector<unsigned int>& elements);
+           std::vector<unsigned short>& elements);
 
       void draw(const Primitives::Matrix& projection, const World::Camera& camera, const Primitives::Matrix& model) const;
