+++ include/apsis/sprite/sheet.h
@@ -167,7 +167,7 @@ namespace Apsis {
       Primitives::VertexArray _vao;
       float* _vertices;
       Primitives::VertexBuffer _vbo;
-      unsigned int* _elements;
+      unsigned short* _elements;
       Primitives::VertexBuffer _ebo;
     };
   }
