+++ include/apsis/world/terrain.h
@@ -35,7 +35,7 @@ namespace Apsis {
       float* _vertices;
       Primitives::VertexBuffer _vbo;
 
-      unsigned int* _elements;
+      unsigned short* _elements;
       Primitives::VertexBuffer _ebo;
 
       std::vector<Apsis::Model::Material> _materials;
